+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 1.0.0
+## 1.0.1
+- Added floating return button in the output autodoc to return readers to the top of the index.
+- Added optional version subtext to the autodoc.
+- Added optional boolean to show or hide timestamp generation.
 
+## 1.0.0
 - Initial version.